editors (can view db name and tables, view and edit rows)

+++ resources/views/livewire/tables.blade.php
@@ -1,29 +1,31 @@
 <div>
     {{-- If your happiness depends on money, you will never be happy with yourself. --}}
 
+    @if($isOwned)
 
-    <div class="py-12">
-        <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
-            <div class="bg-white overflow-hidden shadow-xl sm:rounded-lg">
-                <form wire:submit.prevent="addTable" class="flex flex-col items-center w-full mb-4 md:flex-row md:px-16">
-                    <input
-                        wire:model="tableName"
-                        placeholder="name"
-                        required=""
-                        type="text"
-                        class="flex-grow w-full h-12 px-4 mb-3 transition duration-200 bg-white border border-gray-300 rounded shadow-sm appearance-none md:mr-2 md:mb-0 focus:border-deep-purple-accent-400 focus:outline-none focus:shadow-outline"
-                    />
-                    <button
+        <div class="py-12">
+            <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
+                <div class="bg-white overflow-hidden shadow-xl sm:rounded-lg">
+                    <form wire:submit.prevent="addTable" class="flex flex-col items-center w-full mb-4 md:flex-row md:px-16">
+                        <input
+                            wire:model="tableName"
+                            placeholder="name"
+                            required=""
+                            type="text"
+                            class="flex-grow w-full h-12 px-4 mb-3 transition duration-200 bg-white border border-gray-300 rounded shadow-sm appearance-none md:mr-2 md:mb-0 focus:border-deep-purple-accent-400 focus:outline-none focus:shadow-outline"
+                        />
+                        <button
 
-                        type="submit"
-                        class="inline-flex items-center justify-center w-full h-12 px-6 font-medium tracking-wide transition duration-200 rounded shadow-md md:w-auto bg-deep-purple-accent-400 hover:bg-deep-purple-accent-700 focus:shadow-outline focus:outline-none"
-                    >
-                        Add Table
-                    </button>
-                </form>
+                            type="submit"
+                            class="inline-flex items-center justify-center w-full h-12 px-6 font-medium tracking-wide transition duration-200 rounded shadow-md md:w-auto bg-deep-purple-accent-400 hover:bg-deep-purple-accent-700 focus:shadow-outline focus:outline-none"
+                        >
+                            Add Table
+                        </button>
+                    </form>
+                </div>
             </div>
         </div>
-    </div>
+    @endif
 
     <div class="py-12">
         <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
@@ -45,8 +47,10 @@
 
                             <td class="py-4 px-6 border-b border-grey-light">
                                 <button wire:click="redirectToTableFill({{ $table->id }})" class="text-grey-lighter font-bold py-1 px-3 rounded text-xs bg-green hover:bg-green-dark">fill</button>
-                                <button wire:click="redirectToTableBuild({{ $table->id }})" class="text-grey-lighter font-bold py-1 px-3 rounded text-xs bg-blue hover:bg-blue-dark">build</button>
-                                <button wire:click="deleteTable({{$table}})" class="text-grey-lighter font-bold py-1 px-3 rounded text-xs bg-blue hover:bg-blue-dark">delete</button>
+                                @if($isOwned)
+                                    <button wire:click="redirectToTableBuild({{ $table->id }})" class="text-grey-lighter font-bold py-1 px-3 rounded text-xs bg-blue hover:bg-blue-dark">build</button>
+                                    <button wire:click="deleteTable({{$table}})" class="text-grey-lighter font-bold py-1 px-3 rounded text-xs bg-blue hover:bg-blue-dark">delete</button>
+                                @endif
                             </td>
                         </tr>
                     @endforeach
